Continued packages restructuring
