+++ ccr/__init__.py
@@ -1,4 +1,6 @@
+__version__ = 0.2
 __all__ = ["search", "info", "msearch", "list_orphans",
            "getlatest", "geturl", "getpkgurl", "getpkgbuild",
-           "getpkgbuildraw", "getfileraw", "CCRSession"]
-__version__ = 0.2
+           "getpkgbuildraw", "getfileraw", "Session",
+           "CCR_BASE", "CCR_RPC", "CCR_PKG", "CCR_SUBMIT"
+]
