+++ ccrauth/sqlite.py
@@ -5,7 +5,7 @@ import logging
 from ccrauth.ccrauth import CCRAuth
 
 
-class CCRAuthSQLite(CCRAuth):
+class SQLite(CCRAuth):
 
     def __init__(self):
         """ define username and password if the table auth exists
